+++ test/Chainweb/Test/Rosetta/Server.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 -- |
 -- Module: Chainweb.Test.Rosetta.Server
 -- Copyright: Copyright © 2018 - 2020 Kadena LLC.
@@ -8,80 +10,109 @@
 --
 module Chainweb.Test.Rosetta.Server where
 
-import Control.Error.Util
-import Control.Lens ((^?))
-import Control.Monad (void, foldM)
-import Control.Monad.Except (throwError)
-import Control.Monad.IO.Class
-import Control.Monad.Trans.Class
-import Control.Monad.Trans.Except
 import Data.Aeson
-import Data.Bifunctor
 import Data.Map (Map)
-import Data.Decimal
-import Data.CAS
-import Data.String
-import Data.Proxy (Proxy(..))
-import Data.Tuple.Strict (T2(..))
-import Data.Word (Word64)
 
-import qualified Data.ByteString.Short as BSS
-import qualified Data.DList as DList
-import qualified Data.HashMap.Strict as HM
-import qualified Data.HashSet as HS
 import qualified Data.Map as M
-import qualified Data.Memory.Endian as BA
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
 import qualified Data.Vector as V
 
-import Numeric.Natural
-
-import Pact.Types.ChainMeta (PublicMeta(..))
 import Pact.Types.Command
 import Pact.Types.Hash
-import Pact.Types.Runtime (TxId(..), TxLog(..), Domain(..))
-import Pact.Types.RPC
-import Pact.Types.PactValue (PactValue(..))
-import Pact.Types.Pretty (renderCompactText)
-import Pact.Types.Exp (Literal(..))
-import Pact.Types.Util (fromText')
+import Pact.Types.Runtime (TxId(..))
 
 import Rosetta
 
-import Servant.API
-import Servant.Server
-
 -- internal modules
 
-import Chainweb.BlockCreationTime (BlockCreationTime(..))
-import Chainweb.BlockHash
-import Chainweb.BlockHeader (BlockHeader(..))
-import Chainweb.BlockHeader.Genesis (genesisBlockHeader)
-import Chainweb.BlockHeight (BlockHeight(..))
-import Chainweb.Chainweb.ChainResources (ChainResources(..))
-import Chainweb.Cut
-import Chainweb.CutDB
-import Chainweb.HostAddress
-import Chainweb.Mempool.Mempool
-import Chainweb.Pact.RestAPI.Server
-import Chainweb.Pact.Templates
-import Chainweb.Pact.Service.Types (Domain'(..), BlockTxHistory(..))
-import Chainweb.Payload hiding (Transaction(..))
-import Chainweb.Payload.PayloadStore
-import qualified Chainweb.RestAPI.NetworkID as ChainwebNetId
-import Chainweb.RestAPI.Utils
 import Chainweb.Rosetta.RestAPI
-import Chainweb.Time
-import Chainweb.Transaction (ChainwebTransaction)
-import Chainweb.TreeDB (seekAncestor)
-import Chainweb.Utils
-import Chainweb.Utils.Paging
-import Chainweb.Version
-import Chainweb.WebPactExecutionService (PactExecutionService(..))
-
-import P2P.Node.PeerDB
-import P2P.Node.RestAPI.Server (peerGetHandler)
-import P2P.Peer
+import Chainweb.Rosetta.RestAPI.Server
+import Chainweb.Rosetta.Util
 
 ---
+
+-- TODO: Incorporate into unit tests
+type MockTxResult = (Maybe TxId, T.Text)
+mockTxLogs :: (Map TxId [AccountLog], MockTxResult, V.Vector MockTxResult)
+mockTxLogs = (logs, initial, rest)
+  where
+    (log1,initial) =
+      let key = "miner1"
+          amt = 2.0
+          g = toJSON (key <> "PublicKey" :: T.Text)
+          tid = TxId 1
+          l = [(key, amt, g)]
+          a = (Just tid, "ReqKey1")
+      in ((tid,l), a)
+
+    -- successful, non-coin contract tx
+    (logs2,tx1) =
+      let minerKey = "miner1"
+          key = "sender1"
+          gMiner = toJSON (key <> "PublicKey" :: T.Text)
+          gKey = toJSON (key <> "PublicKey" :: T.Text)
+          (fundTid, tid, gasTid) = (TxId 2, TxId 3, TxId 4)
+          fundLogs = (fundTid, [(key, 10.0, gKey)])
+          gasLogs = (gasTid, [(minerKey, 12.0, gMiner)])
+          a = (Just tid, "ReqKey2")
+      in ([fundLogs,gasLogs], a)
+
+    (logs3,tx2) =
+      let minerKey = "miner1"
+          key = "sender1"
+          gMiner = toJSON (key <> "PublicKey" :: T.Text)
+          gKey = toJSON (key <> "PublicKey" :: T.Text)
+          (fundTid, tid, gasTid) = (TxId 5, TxId 6, TxId 7)
+          fundLogs = (fundTid, [(key, 10.0, gKey)])
+          gasLogs = (gasTid, [(minerKey, 12.0, gMiner)])
+          a = (Just tid, "ReqKey3")
+      in ([fundLogs,gasLogs], a)
+
+    -- successful, coin contract tx
+    (logs4,tx3) =
+      let minerKey = "miner1"
+          key = "sender1"
+          gMiner = toJSON (key <> "PublicKey" :: T.Text)
+          gKey = toJSON (key <> "PublicKey" :: T.Text)
+          (fundTid, tid, gasTid) = (TxId 8, TxId 9, TxId 10)
+          fundLogs = (fundTid, [(key, 10.0, gKey)])
+          transferLogs = (tid, [(key, 5.0, gKey)])
+          gasLogs = (gasTid, [(minerKey, 12.0, gMiner)])
+          a = (Just tid, "ReqKey4")
+      in ([fundLogs,transferLogs,gasLogs], a)
+
+    -- unsuccessful tx
+    (logs5,tx4) =
+      let minerKey = "miner1"
+          key = "sender1"
+          gMiner = toJSON (key <> "PublicKey" :: T.Text)
+          gKey = toJSON (key <> "PublicKey" :: T.Text)
+          (fundTid, gasTid) = (TxId 11, TxId 12)
+          fundLogs = (fundTid, [(key, 10.0, gKey)])
+          gasLogs = (gasTid, [(minerKey, 12.0, gMiner)])
+          a = (Nothing, "ReqKey5")
+      in ([fundLogs,gasLogs], a)
+
+    rest = V.fromList [tx1, tx2, tx3, tx4]
+    logs = M.fromList $ [log1] <> logs2 <> logs3 <> logs4 <> logs5
+
+testBlock :: Either String [(T.Text, [Operation])]
+testBlock = nonGenesisTransactions logs getTxId toTx initial rest
+  where
+    toTx (_, rk) ops = (rk, ops)
+    getTxId (tid, _) = tid
+    (logs, initial, rest) = mockTxLogs
+
+testBlockTransaction :: T.Text -> Either RosettaFailure (T.Text, [Operation])
+testBlockTransaction trk =
+  nonGenesisTransaction logs toRk getTxId toTx initial rest (RequestKey $ Hash $ T.encodeUtf8 trk)
+  where
+    toTx (_, rk) ops = (rk, ops)
+    toRk (_, rk) = RequestKey $ Hash $ T.encodeUtf8 rk
+    getTxId (tid, _) = tid
+    (logs, initial, rest) = mockTxLogs
+
+-- TODO: test for `validateNetwork`
+
+-- TODO: test for `kdaToRosettaAmount`
